requirements.txt
pip~=22.3.1
attrs~=23.2.0
wheel~=0.38.4
pillow~=10.4.0
docutils~=0.18.1
Sphinx~=5.0.2
jinja2~=3.1.4
pandas~=2.1.4
selenium~=4.23.1
selenium-stealth
alpaca-py=0.28.0

